# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Recal API
  version: 0.1.0
  description: API backend de l’agent autonome de veille d’opportunités.
servers:
  - url: http://localhost:8000
    description: Développement local
  - url: https://api.example.com
    description: Production — à remplacer

tags:
  - name: Health
  - name: Profile
  - name: Opportunities
  - name: Runs

paths:
  /api/v1/health:
    get:
      tags: [Health]
      summary: Vérifier la disponibilité du backend
      operationId: getHealth
      responses:
        '200':
          description: Backend disponible
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /api/v1/profile:
    get:
      tags: [Profile]
      summary: Lire le profil courant
      operationId: getProfile
      responses:
        '200':
          description: Profil utilisateur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
    put:
      tags: [Profile]
      summary: Remplacer le profil courant
      operationId: updateProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfileUpdate'
      responses:
        '200':
          description: Profil mis à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '422':
          $ref: '#/components/responses/ValidationError'

  /api/v1/opportunities:
    get:
      tags: [Opportunities]
      summary: Lister les opportunités
      operationId: listOpportunities
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: status
          in: query
          schema:
            type: string
            enum: [new, seen, saved, dismissed, expired]
        - name: min_score
          in: query
          schema:
            type: number
            minimum: 0
            maximum: 100
      responses:
        '200':
          description: Liste paginée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpportunityPage'

  /api/v1/opportunities/{opportunity_id}:
    get:
      tags: [Opportunities]
      summary: Lire une opportunité
      operationId: getOpportunity
      parameters:
        - $ref: '#/components/parameters/OpportunityId'
      responses:
        '200':
          description: Opportunité trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Opportunity'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/v1/opportunities/{opportunity_id}/feedback:
    post:
      tags: [Opportunities]
      summary: Enregistrer le feedback utilisateur
      operationId: submitOpportunityFeedback
      parameters:
        - $ref: '#/components/parameters/OpportunityId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackRequest'
      responses:
        '200':
          description: Feedback enregistré
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/v1/runs:
    post:
      tags: [Runs]
      summary: Déclencher un cycle de veille
      description: Réservé au développement et à l’administration dans le MVP initial.
      operationId: createWatchRun
      responses:
        '202':
          description: Cycle accepté
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Run'

  /api/v1/runs/{run_id}:
    get:
      tags: [Runs]
      summary: Consulter le statut d’un cycle
      operationId: getWatchRun
      parameters:
        - name: run_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Cycle trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Run'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  parameters:
    OpportunityId:
      name: opportunity_id
      in: path
      required: true
      schema:
        type: string
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20

  responses:
    NotFound:
      description: Ressource introuvable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Données invalides
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  schemas:
    HealthResponse:
      type: object
      required: [status, service, version]
      properties:
        status:
          type: string
          example: ok
        service:
          type: string
          example: recal-api
        version:
          type: string
          example: 0.1.0

    UserProfile:
      type: object
      required: [id, interests, countries, study_level, skills, relevance_threshold]
      properties:
        id:
          type: string
        interests:
          type: array
          items: {type: string}
        countries:
          type: array
          items: {type: string}
        study_level:
          type: string
        skills:
          type: array
          items: {type: string}
        relevance_threshold:
          type: number
          minimum: 0
          maximum: 100

    UserProfileUpdate:
      type: object
      required: [interests, countries, study_level, skills, relevance_threshold]
      properties:
        interests:
          type: array
          items: {type: string}
        countries:
          type: array
          items: {type: string}
        study_level:
          type: string
        skills:
          type: array
          items: {type: string}
        relevance_threshold:
          type: number
          minimum: 0
          maximum: 100

    Opportunity:
      type: object
      required: [id, type, title, summary, relevance_score, confidence_score, source_url, status]
      properties:
        id: {type: string}
        type:
          type: string
          enum: [hackathon, internship, fellowship, scholarship, other]
        title: {type: string}
        organization: {type: string}
        summary: {type: string}
        relevance_score:
          type: number
          minimum: 0
          maximum: 100
        confidence_score:
          type: number
          minimum: 0
          maximum: 100
        relevance_reasons:
          type: array
          items: {type: string}
        deadline:
          type: string
          format: date
          nullable: true
        eligibility:
          type: object
          additionalProperties: true
        source_url:
          type: string
          format: uri
        verified_at:
          type: string
          format: date-time
        status:
          type: string
          enum: [new, seen, saved, dismissed, expired]

    OpportunityPage:
      type: object
      required: [items, page, page_size, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Opportunity'
        page: {type: integer}
        page_size: {type: integer}
        total: {type: integer}

    FeedbackRequest:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [interested, not_interested, saved, dismissed]
        comment:
          type: string
          nullable: true

    FeedbackResponse:
      type: object
      required: [opportunity_id, action, recorded_at]
      properties:
        opportunity_id: {type: string}
        action: {type: string}
        recorded_at:
          type: string
          format: date-time

    Run:
      type: object
      required: [id, status, created_at]
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum: [queued, running, completed, failed]
        created_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
        opportunities_found:
          type: integer
          minimum: 0

    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code: {type: string}
        message: {type: string}
        details:
          type: object
          additionalProperties: true